feat: inzichten suspense
